requirements.txt
allure-pytest==2.15.0
python-dotenv==1.1.1
pytest==8.4.2
pytest-xdist==3.6.1
pytest-rerunfailures==14.0
playwright==1.55.0
faker==37.11.0
Pillow==12.0.0
